upgrade hardhat-diamond-abi to 3.0.1
