+++ .vscode/launch.json
@@ -32,15 +32,15 @@
             "name": "Multichain Plugin Test",
             "type": "node",
             "request": "launch",
-            // "preLaunchTask": "tsc: build - tsconfig.json",
+            "preLaunchTask": "tsc: build - tsconfig.json",
             "program": "${workspaceFolder}/node_modules/hardhat/internal/cli/cli.js",
             "args": [ 
                 "test-multichain", 
-                // "${workspaceFolder}/test/integration/multichain/tests/MultiChainForkDeploy.test.ts",
-                // "${workspaceFolder}/test/integration/multichain/tests/NFTFactory.test.ts",
-                // "${workspaceFolder}/test/integration/multichain/tests/GNUSBridge.test.ts",
-                "${workspaceFolder}/test/integration/multichain/tests/GNUSERC20.test.ts",             
-                // "${workspaceFolder}/test/integration/multichain/tests/Erc20Batch.test.ts",
+                "${workspaceFolder}/test/deployment/MultiChainForkDeploy.test.ts",
+                "${workspaceFolder}/test/unit/NFTFactory.test.ts",
+                "${workspaceFolder}/test/unit/GNUSBridge.test.ts",
+                "${workspaceFolder}/test/unit/GNUSERC20.test.ts",             
+                "${workspaceFolder}/test/unit/Erc20Batch.test.ts",
                 // "--chains", "hardhat", "--logs", "logs",
                 // "--chains", "sepolia", "--logs", "logs",
                 // "--chains", "polygon_amoy", "--logs", "logs",
@@ -50,7 +50,7 @@
                 // "--chains", "base", "--logs", "logs",
                 // "--chains", "bsc_testnet", "--logs", "logs",
                 // "--chains", "base_sepolia", "--logs", "logs",
-                // "--chains", "hardhat,sepolia", "--logs", "logs",
+                "--chains", "hardhat,sepolia", "--logs", "logs",
                 // "--chains", "sepolia,polygon_amoy", "--logs", "logs",
                 // "--chains", "hardhat,sepolia,polygon_amoy", "--logs", "logs", 
             ],
@@ -79,12 +79,12 @@
             "program": "${workspaceFolder}/node_modules/hardhat/internal/cli/cli.js",
             "args": [ 
                 "test-multichain", 
-                // "./test/integration/multichain/tests/*",
-                "${workspaceFolder}/test/integration/multichain/tests/NFTFactoryTests.ts",
-                "${workspaceFolder}/test/integration/multichain/tests/MultiChainForkDeployTests.ts",
-                "${workspaceFolder}/test/integration/multichain/tests/GNUSBridgeTests.ts",
+                // "./test/unit/*",
+                "${workspaceFolder}/test/unit/NFTFactoryTests.ts",
+                "${workspaceFolder}/test/unit/MultiChainForkDeployTests.ts",
+                "${workspaceFolder}/test/unit/GNUSBridgeTests.ts",
                 "${workspaceFolder}/multichain/tests/GNUSERC20Tests.ts",             
-                "${workspaceFolder}/test/integration/multichain/tests/Erc20BatchTests.ts",
+                "${workspaceFolder}/test/unit/Erc20BatchTests.ts",
                 "--chains", "polygon_amoy,sepolia", "--logs", "logs"],
             "runtimeArgs": [ "--preserve-symlinks" ],
             "skipFiles": [
